use gradle vs. mvn to build ethereumj
This reflects changes made in ethereum/ethereumj#179, porting
ethereumj's build system from Maven to Gradle. It should fix the build
failure seen at
http://build.ethdev.com/builders/Linux%20EthereumJ%20PRs/builds/15.

Note that the command being used (`./gradlew`) is a script residing in
the root of the ethereumj repository. This means that there is no need
to install Gradle on the host machine/VM.

For complete information on Gradle and the `gradlew` script see
http://www.gradle.org and
http://www.gradle.org/docs/current/userguide/gradle_wrapper.html.